copied code from draftjs component and put it directly into modals on uploadvideo.js
